refactor(cli): split commands into files
