+++ pyproject.toml
@@ -8,6 +8,7 @@ requires-python = ">=3.12"
 
 dependencies = [
     "google-genai>=1.4.0",
+    "langchain>=0.3.20",
     "pydantic>=2.10.6",
     "typer>=0.15.2",
 ]
@@ -29,7 +30,7 @@ dev = [
 ]
 
 [project.scripts]
-grim = "grimoire.cli:cli"
+grim = "grimoire.main:cli"
 
 [project.urls]
 repository = "https://github.com/BDP25/grimoire"
